Split slashing params & signing info into separate files
